Remove IgnoreExitCode on inner build steps for compilation and generation of configuration files.
The IgnoreExitCode appears to have been added because RoboCopy returns a non-zero exit code on success (as opposed to most DOS-legacy executables which only return non-zero to indicate a problem).  However, this is not necessary as the RoboCopy build task handles the exit code internally, and only returns an error if the exit code is one of error values.

REF: https://github.com/loresoft/msbuildtasks/blob/master/Source/MSBuild.Community.Tasks/RoboCopy.cs#L443
REF: https://msdn.microsoft.com/en-us/library/microsoft.build.utilities.tooltask.handletaskexecutionerrors(v=vs.90).aspx

git-svn-id: https://ncisvn.nci.nih.gov/svn/oce_wcmteam/ContentDeliveryEngine/branches/git-build@19962 e9cc8610-1bbf-714c-a5f3-26df0ede829a