[NCISEO-11]:
- After talking with Michael and reviewing my code, he offered the following suggestions to refactor and make the code tighter. I have refactored and implemented most of the suggestions below. I will review the code again on Monday to make sure that it does have all the suggestions made by Bryan

I) DynamicTrialListingControl
Onload:

Remove commented out Session references and never use session
OnEmptyResults:
Instead of splitting parameters you should ask the derived class for an array of the parameters
Also there is a story in this sprint for friendly names, so the actual parameters that we need to use in the redirect may not be what is in the URL
Use NCIUrl for manipulating the URL.  (Less string manipulation)
Response.Redirect should pass in additional boolean to stop page processing
Never catch a Response.Redirect — ever.

II) DynamicTrialListingPageDiseaseControl

ParseURL:

Why are you checking the URL for no trials when the base class could do that?  Then you could say:
If this.IsNoTrialsDisplay() then SetParametersForNoTrials
The code should be refactored to handle when it is no trial page and when it is regular — but only to get the “array” of parameters.
I.E. the work like this.InterventionIDs = string.join is being duplicated and if we change how that logic works, then we have 2 places to change
SetParametersForNoTrialsPage:
Use NCIUrl for parsing query params, not current context or string checks.

III) Intervention Listing
ParseURL:
How do you handle notrials??​