Triple-quote the source path parameter
Passing strings to PowerShell is painful.

If the string (e.g. a path) has spaces, you need to quote it, right?
Ha! The command interpreter eats those.  And the string still doesn't
get passed correctly.

Instead, you have to *triple* quote it.
So -fancyParam "'string with spaces.'"

And because that's easy to read:
quotation mark - apostrophe - actual value - closing apostrophe - closing quotation mark.

That could be "difficult" if you ever needed to pass a string containing
an apostrophe.